added some modularization
