+++ SPhangman.mjs
@@ -1,0 +1,3 @@
+import * as hangman from "./lib/hangman.mjs";
+
+hangman.init()
