+++ SinglePlayer.html
@@ -81,6 +81,6 @@
         </div>
     </section>
 
-    <script src="hangman.js"></script>
+    <script src="SPhangman.mjs" type="module"></script>
 </body>
 </html>
